+++ packages/grid/README.md
@@ -1,0 +1,6 @@
+<a href="https://nivo.rocks"><img alt="nivo" src="https://raw.githubusercontent.com/plouc/nivo/master/nivo.png" width="216" height="68"/></a>
+
+# `@nivo/legends`
+
+[![version](https://img.shields.io/npm/v/@nivo/legends?style=for-the-badge)](https://www.npmjs.com/package/@nivo/legends)
+[![downloads](https://img.shields.io/npm/dm/@nivo/legends?style=for-the-badge)](https://www.npmjs.com/package/@nivo/legends)
