+++ packages/grid/package.json
@@ -1,0 +1,27 @@
+{
+    "name": "@nivo/grid",
+    "description": "grid components for nivo dataviz library",
+    "version": "0.82.0",
+    "license": "MIT",
+    "author": {
+        "name": "Raphaël Benitte",
+        "url": "https://github.com/plouc"
+    },
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/plouc/nivo.git",
+        "directory": "packages/grid"
+    },
+    "main": "./dist/nivo-grid.cjs.js",
+    "module": "./dist/nivo-grid.es.js",
+    "types": "./dist/types/index.d.ts",
+    "files": [
+        "README.md",
+        "LICENSE.md",
+        "dist/",
+        "!dist/tsconfig.tsbuildinfo"
+    ],
+    "publishConfig": {
+        "access": "public"
+    }
+}
