+++ packages/grid/src/types.ts
@@ -1,0 +1,8 @@
+export type Vertex = [number, number]
+
+export interface BoxEdges {
+    top: number
+    right: number
+    bottom: number
+    left: number
+}
