+++ packages/waffle/package.json
@@ -29,8 +29,9 @@
     ],
     "dependencies": {
         "@nivo/colors": "workspace:*",
+        "@nivo/core": "workspace:*",
+        "@nivo/grid": "workspace:*",
         "@nivo/legends": "workspace:*",
-        "@nivo/recompose": "workspace:*",
         "@nivo/tooltip": "workspace:*",
         "@types/d3-shape": "^2.0.0",
         "@types/prop-types": "^15.7.2",
@@ -39,11 +40,7 @@
         "@react-spring/web": "9.4.5 || ^9.7.2",
         "d3-shape": "^1.3.5"
     },
-    "devDependencies": {
-        "@nivo/core": "workspace:*"
-    },
     "peerDependencies": {
-        "@nivo/core": "0.82.1",
         "react": ">= 16.14.0 < 19.0.0"
     },
     "publishConfig": {
